separated environment from interpreter
environment is a standalone part of the interpreter, as opposed to be
fully integrated. Ideally, the interpreter could use alternative
environments, however, the code is not so flexible at this point...

environment.py: the new interface should make it easier to add unittests
for the compiler. A few of them have been implemented (see tests/), but
the majority of the work still has to be done. Testing built-in
functions requires specific test cases in order to make sure that
everything behaves as it should.

setup.py: now requires the 'peppercorn' package, as opposed to the
'enumerator' package.