NusDomain and NusCDomain
...replace the old Domain class. They inherig from DNAObjects in order
to reduce the code-base and bring the packages closer together.
However, there are some things in DNAObjects that I am not happy with,
for example the strict watson-crick complementarity of domains, as well
as the strict 1:1 mapping between d1 and d1*.

NusDomain requires the user to specify sequence constraints for
top/bottom strands, usually 'HHHHH' and 'DDDDD', and to specify a
tag (e.g. 'toehold') for sequence design.

srinivas2015.ts shows how to address domain-level design choices within
the translation-scheme.